chore: Integer -> Long 변경
의미없는 형변환 제거를 위한 변경

+++ BE-MyCarMaster/src/main/java/softeer/bemycarmaster/api/domain/trim/dto/request/GetTrimsRequest.java
@@ -9,5 +9,5 @@ import lombok.Setter;
 public class GetTrimsRequest {
 
 	@Schema(description = "모델 식별자", example = "1")
-	private Integer modelId;
+	private Long modelId;
 }
